requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
sqlparse==0.5.3
gunicorn==20.1.0  # для продакшн-сервера
django-crispy-forms==1.14.0  # если используешь crispy forms
